+++ src/service-worker.js
@@ -1,5 +1,13 @@
 /* eslint-disable no-undef, no-restricted-globals */
 
+self.addEventListener("install", () => {
+    self.skipWaiting();
+});
+
+self.addEventListener("activate", () => {
+    return self.clients.claim();
+});
+
 // Source: https://dev.to/drbragg/handling-service-worker-updates-in-your-vue-pwa-1pip
 self.addEventListener("message", event => {
     if (event.data && event.data.type === "SKIP_WAITING") {